+++ src/main/java/module-info.java
@@ -17,4 +17,6 @@ module com.example.addressbook {
     opens com.example.addressbook.controller to javafx.fxml;
     exports com.example.addressbook.model;
     opens com.example.addressbook.model to javafx.fxml;
+
+
 }
